optimize audit log

+++ server/src/main/java/org/apache/iotdb/db/service/thrift/impl/TSServiceImpl.java
@@ -224,9 +224,10 @@ public class TSServiceImpl implements TSIService.Iface {
       plan.setLoginUserName(username);
 
       QUERY_FREQUENCY_RECORDER.incrementAndGet();
-      AuditLogUtils.writeAuditLog(
-          String.format("Session %s execute Query: %s", session, statement));
-
+      if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+        AuditLogUtils.writeAuditLog(
+            String.format("Session %s execute Query: %s", session, statement));
+      }
       final long queryId = SESSION_MANAGER.requestQueryId(statementId, true);
       QueryContext context =
           serviceProvider.genQueryContext(
@@ -646,7 +647,10 @@ public class TSServiceImpl implements TSIService.Iface {
           serviceProvider
               .getPlanner()
               .parseSQLToPhysicalPlan(statement, session.getZoneId(), session.getClientVersion());
-      AuditLogUtils.writeAuditLog(req.getStatement());
+      if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+        AuditLogUtils.writeAuditLog(req.getStatement());
+      }
+
       if (physicalPlan.isQuery()) {
         return submitQueryTask(session, physicalPlan, startTime, req);
       } else {
@@ -678,7 +682,9 @@ public class TSServiceImpl implements TSIService.Iface {
           serviceProvider
               .getPlanner()
               .parseSQLToPhysicalPlan(statement, session.getZoneId(), session.getClientVersion());
-      AuditLogUtils.writeAuditLog(statement);
+      if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+        AuditLogUtils.writeAuditLog(statement);
+      }
       if (physicalPlan.isQuery()) {
         return submitQueryTask(session, physicalPlan, startTime, req);
       } else {
@@ -960,8 +966,10 @@ public class TSServiceImpl implements TSIService.Iface {
     final QueryPlan queryPlan = selectIntoPlan.getQueryPlan();
 
     QUERY_FREQUENCY_RECORDER.incrementAndGet();
-    AuditLogUtils.writeAuditLog(
-        String.format("Session %s execute select into: %s", session, statement));
+    if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+      AuditLogUtils.writeAuditLog(
+          String.format("Session %s execute select into: %s", session, statement));
+    }
 
     if (queryPlan.isEnableTracing()) {
       TRACING_MANAGER.setSeriesPathNum(queryId, queryPlan.getPaths().size());
@@ -1214,7 +1222,8 @@ public class TSServiceImpl implements TSIService.Iface {
       return loginStatus;
     }
 
-    if (conf.isEnableAuditLogWrite()) {
+    if (conf.isEnableAuditLogWrite()
+        && !AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
       AuditLogUtils.writeAuditLog(
           String.format(
               "Session %s insertRecords, first device %s, first time %s",
@@ -1303,7 +1312,8 @@ public class TSServiceImpl implements TSIService.Iface {
       return loginStatus;
     }
 
-    if (conf.isEnableAuditLogWrite()) {
+    if (conf.isEnableAuditLogWrite()
+        && !AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
       AuditLogUtils.writeAuditLog(
           String.format(
               "Session %s insertRecords, device %s, first time %s",
@@ -1350,7 +1360,8 @@ public class TSServiceImpl implements TSIService.Iface {
       return loginStatus;
     }
 
-    if (conf.isEnableAuditLogWrite()) {
+    if (conf.isEnableAuditLogWrite()
+        && !AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
       AuditLogUtils.writeAuditLog(
           String.format(
               "Session %s insertRecords, device %s, first time %s",
@@ -1409,7 +1420,8 @@ public class TSServiceImpl implements TSIService.Iface {
       return loginStatus;
     }
 
-    if (conf.isEnableAuditLogWrite()) {
+    if (conf.isEnableAuditLogWrite()
+        && !AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
       AuditLogUtils.writeAuditLog(
           String.format(
               "Session %s insertRecords, first device %s, first time %s",
@@ -1523,7 +1535,9 @@ public class TSServiceImpl implements TSIService.Iface {
       if (isStatusNotSuccess(loginStatus)) {
         return loginStatus;
       }
-      if (!req.getPrefixPath().startsWith(SYSTEM_STORAGE_GROUP) && conf.isEnableAuditLogWrite()) {
+      if (!req.getPrefixPath().startsWith(SYSTEM_STORAGE_GROUP)
+          && conf.isEnableAuditLogWrite()
+          && !AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
         AuditLogUtils.writeAuditLog(
             String.format(
                 "Session %s insertRecord, device %s, time %s",
@@ -1560,7 +1574,8 @@ public class TSServiceImpl implements TSIService.Iface {
       if (isStatusNotSuccess(loginStatus)) {
         return loginStatus;
       }
-      if (conf.isEnableAuditLogWrite()) {
+      if (conf.isEnableAuditLogWrite()
+          && !AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
         AuditLogUtils.writeAuditLog(
             String.format(
                 "Session %s insertRecord, device %s, time %s",
@@ -1771,9 +1786,10 @@ public class TSServiceImpl implements TSIService.Iface {
       if (isStatusNotSuccess(loginStatus)) {
         return loginStatus;
       }
-
-      AuditLogUtils.writeAuditLog(
-          String.format("Session-%s create timeseries %s", session, req.getPath()));
+      if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+        AuditLogUtils.writeAuditLog(
+            String.format("Session-%s create timeseries %s", session, req.getPath()));
+      }
 
       CreateTimeSeriesPlan plan =
           new CreateTimeSeriesPlan(
@@ -1803,11 +1819,12 @@ public class TSServiceImpl implements TSIService.Iface {
       if (isStatusNotSuccess(loginStatus)) {
         return loginStatus;
       }
-
-      AuditLogUtils.writeAuditLog(
-          String.format(
-              "Session-%s create aligned timeseries %s.%s",
-              session, req.getPrefixPath(), req.getMeasurements()));
+      if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+        AuditLogUtils.writeAuditLog(
+            String.format(
+                "Session-%s create aligned timeseries %s.%s",
+                session, req.getPrefixPath(), req.getMeasurements()));
+      }
 
       List<TSDataType> dataTypes = new ArrayList<>();
       for (int dataType : req.dataTypes) {
@@ -1849,10 +1866,12 @@ public class TSServiceImpl implements TSIService.Iface {
       if (isStatusNotSuccess(loginStatus)) {
         return loginStatus;
       }
-      AuditLogUtils.writeAuditLog(
-          String.format(
-              "Session-%s create %s timeseries, the first is %s",
-              session, req.getPaths().size(), req.getPaths().get(0)));
+      if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+        AuditLogUtils.writeAuditLog(
+            String.format(
+                "Session-%s create %s timeseries, the first is %s",
+                session, req.getPaths().size(), req.getPaths().get(0)));
+      }
 
       CreateMultiTimeSeriesPlan multiPlan = new CreateMultiTimeSeriesPlan();
       List<PartialPath> paths = new ArrayList<>(req.paths.size());
@@ -1964,8 +1983,10 @@ public class TSServiceImpl implements TSIService.Iface {
       if (isStatusNotSuccess(loginStatus)) {
         return loginStatus;
       }
-      AuditLogUtils.writeAuditLog(
-          String.format("Session-%s create schema template %s", session, req.getName()));
+      if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+        AuditLogUtils.writeAuditLog(
+            String.format("Session-%s create schema template %s", session, req.getName()));
+      }
 
       CreateTemplatePlan plan;
       // Construct plan from serialized request
@@ -2068,10 +2089,12 @@ public class TSServiceImpl implements TSIService.Iface {
     if (isStatusNotSuccess(loginStatus)) {
       return loginStatus;
     }
-    AuditLogUtils.writeAuditLog(
-        String.format(
-            "Session-%s set device template %s.%s",
-            session, req.getTemplateName(), req.getPrefixPath()));
+    if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+      AuditLogUtils.writeAuditLog(
+          String.format(
+              "Session-%s set device template %s.%s",
+              session, req.getTemplateName(), req.getPrefixPath()));
+    }
 
     try {
       SetTemplatePlan plan = new SetTemplatePlan(req.templateName, req.prefixPath);
@@ -2089,10 +2112,12 @@ public class TSServiceImpl implements TSIService.Iface {
     if (isStatusNotSuccess(loginStatus)) {
       return loginStatus;
     }
-    AuditLogUtils.writeAuditLog(
-        String.format(
-            "Session-%s unset schema template %s.%s",
-            session, req.getPrefixPath(), req.getTemplateName()));
+    if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+      AuditLogUtils.writeAuditLog(
+          String.format(
+              "Session-%s unset schema template %s.%s",
+              session, req.getPrefixPath(), req.getTemplateName()));
+    }
     try {
       UnsetTemplatePlan plan = new UnsetTemplatePlan(req.prefixPath, req.templateName);
       TSStatus status = serviceProvider.checkAuthority(plan, session);
@@ -2110,10 +2135,12 @@ public class TSServiceImpl implements TSIService.Iface {
     if (isStatusNotSuccess(loginStatus)) {
       return loginStatus;
     }
-
-    AuditLogUtils.writeAuditLog(
-        String.format(
-            "Session-%s unset using schema template %s on %s", session, templateName, prefixPath));
+    if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+      AuditLogUtils.writeAuditLog(
+          String.format(
+              "Session-%s unset using schema template %s on %s",
+              session, templateName, prefixPath));
+    }
 
     try {
       DeactivateTemplatePlan plan =
@@ -2132,10 +2159,12 @@ public class TSServiceImpl implements TSIService.Iface {
     if (isStatusNotSuccess(loginStatus)) {
       return loginStatus;
     }
-    AuditLogUtils.writeAuditLog(
-        String.format(
-            "Session-%s create timeseries of schema template on path %s",
-            session, req.getDstPath()));
+    if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+      AuditLogUtils.writeAuditLog(
+          String.format(
+              "Session-%s create timeseries of schema template on path %s",
+              session, req.getDstPath()));
+    }
 
     try {
       ActivateTemplatePlan plan = new ActivateTemplatePlan(new PartialPath(req.getDstPath()));
@@ -2153,8 +2182,10 @@ public class TSServiceImpl implements TSIService.Iface {
     if (isStatusNotSuccess(loginStatus)) {
       return loginStatus;
     }
-    AuditLogUtils.writeAuditLog(
-        String.format("Session-%s drop schema template %s.", session, req.getTemplateName()));
+    if (!AuditLogUtils.LOG_LEVEL_NONE.equals(conf.getAuditLogStorage())) {
+      AuditLogUtils.writeAuditLog(
+          String.format("Session-%s drop schema template %s.", session, req.getTemplateName()));
+    }
 
     DropTemplatePlan plan = new DropTemplatePlan(req.templateName);
     TSStatus status = serviceProvider.checkAuthority(plan, session);
